openAPI/requirements.txt
#這個requirements.txt是羅列在 這個 openAPI 專案下. 在Git Bash終端機下(python3.10.8)下
#**起初是在    JZ@WINDOWS-QET06QS MINGW64 ~/Documents/GitHub/2024_01_08_python_openAPI(main)下
#**先執行指令: cd openAPI
#**變成        JZ@WINDOWS-QET06QS MINGW64 ~/Documents/GitHub/2024_01_08_python_openAPI/openAPI (main)
#指令: python -m venv .venv             ➤➤➤➤建立虛擬虛擬主機 .venv
#指令: source .venv/scripts/activate    ➤➤➤➤進入虛擬主機 .venv  
#指令: pip install -r requirements.txt  ➤➤➤➤安裝的套件的列表
#
#
#什麼是 Uvicorn？ ◀◀◀◀這個模組在 requirements.txt中有先安裝給虛擬主機了
#   uvicorn 是一個輕量級高效的web伺服器框架。 Uvicorn 是基於 uvloop 和 httptools 構建的非常快速的 ASGI 伺服器。
#
#什麼是 FastAPI？ ◀◀◀◀這個模組在 requirements.txt中有先安裝給虛擬主機了
#   FastAPI是在Python中構建RESTful API的現代Web框架
#
#什麼是 python-dotenv？ ◀◀◀◀這個模組在 requirements.txt中有先安裝給虛擬主機了
#   python-dotenv 模組 : 處理帳號、密碼及金鑰…等敏感資訊的存放問題
#   (請參考說明https://kirin.idv.tw/python-dotenv-sensitive-secret-data-storage/)
#
#
fastapi
redis
uvicorn[standard]
python-dotenv